Added alerts. Added search hook
