+++ src/components/common/UI/Loader/Loader.style.ts
@@ -8,11 +8,11 @@ const LoaderWrapper = styled.div`
     display: inline-block;
     position: relative;
     width: 80px;
-    height: 80px;
+    height: 30px;
 
     > div {
       position: absolute;
-      top: 33px;
+      top: 10px;
       width: 13px;
       height: 13px;
       border-radius: 50%;
